# backend_progemi/app/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/users/new_user": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "New User",
        "description": "Create a new user with the provided user ID.",
        "operationId": "new_user_api_users_new_user_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUserInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "409": {
            "description": "User already exists"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/get_user": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User",
        "description": "Retrieve user information based on the provided user ID.",
        "operationId": "get_user_api_users_get_user_get",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUserInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User does not exist"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/delete_user": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Delete User",
        "description": "Delete a user based on the provided user ID.",
        "operationId": "delete_user_api_users_delete_user_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteUserInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/new_project": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "New Project",
        "description": "Create a new project for the given user.",
        "operationId": "new_project_api_users_new_project_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewProjectInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/delete_project": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Delete Project",
        "description": "Delete a project for the given user.",
        "operationId": "delete_project_api_users_delete_project_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteProjectInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/get_all_projects": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Get All Projects",
        "description": "Retrieve all projects name associated with a user.",
        "operationId": "get_all_projects_api_users_get_all_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllProjectsInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array",
                  "title": "Response Get All Projects Api Users Get All Projects Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/upload_proposal": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Upload Proposal",
        "description": "Upload a proposal file for the given user & project.",
        "operationId": "upload_proposal_api_users_upload_proposal_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_proposal_api_users_upload_proposal_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/get_proposal": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Get Proposal",
        "description": "Return a proposal file for the given user & project.",
        "operationId": "get_proposal_api_users_get_proposal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetProposalInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/delete_proposal": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Delete Proposal",
        "description": "Delete a proposal for the given user & project.",
        "operationId": "delete_proposal_api_users_delete_proposal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteProposalInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/get_all_proposals_infos": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Get All Proposals",
        "description": "Retrieve all proposals associated with a project for a user.",
        "operationId": "get_all_proposals_api_users_get_all_proposals_infos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAllProposalsInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProposalInfos"
                  },
                  "type": "array",
                  "title": "Response Get All Proposals Api Users Get All Proposals Infos Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/get_project_packs": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Project Packs",
        "description": "Retrieve all packs associated with a project for a user.",
        "operationId": "get_project_packs_api_users_get_project_packs_get",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetProjectPacksInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array",
                  "title": "Response Get Project Packs Api Users Get Project Packs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/update_project_packs": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Update Project Packs",
        "description": "Update the packs associated with a project for a user.",
        "operationId": "update_project_packs_api_users_update_project_packs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectPacksInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/get_proposal_extracted_object": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Get Proposal Extracted Object",
        "description": "Retrieve the structured proposal object for a given proposal title.",
        "operationId": "get_proposal_extracted_object_api_users_get_proposal_extracted_object_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetProposalExtractedObjectInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalWithPolygonAndValidation-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/set_proposal_extracted_object": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Set Proposal Extracted Object",
        "description": "Set the structured proposal object for a given proposal title.",
        "operationId": "set_proposal_extracted_object_api_users_set_proposal_extracted_object_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetProposalExtractedObjectInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalWithPolygonAndValidation-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/get_proposal_object": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Get Proposal Object",
        "description": "Retrieve a structured proposal object for a given proposal title.",
        "operationId": "get_proposal_object_api_users_get_proposal_object_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProposalObjectInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalWithPolygonAndValidation-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/create_proposal_object_validation": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create Proposal Object Validation",
        "description": "Retrieve the validation report for a given proposal title.",
        "operationId": "create_proposal_object_validation_api_users_create_proposal_object_validation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProposalObjectValidationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/get_proposal_object_validation": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Proposal Object Validation",
        "description": "Retrieve the validation report for a given proposal title.",
        "operationId": "get_proposal_object_validation_api_users_get_proposal_object_validation_get",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetProposalObjectValidationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_proposal_api_users_upload_proposal_post": {
        "properties": {
          "user_email": {
            "type": "string",
            "format": "email",
            "title": "User Email"
          },
          "project_name": {
            "type": "string",
            "title": "Project Name"
          },
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "user_email",
          "project_name",
          "file"
        ],
        "title": "Body_upload_proposal_api_users_upload_proposal_post"
      },
...
      "DeleteProjectInput": {
        "properties": {
          "user_email": {
            "type": "string",
            "format": "email",
            "title": "User Email",
            "description": "Email of the user"
          },
          "project_name": {
            "type": "string",
            "title": "Project Name",
            "description": "Name of the project to delete"
          }
        },
        "type": "object",
        "required": [
          "user_email",
          "project_name"
        ],
        "title": "DeleteProjectInput",
        "description": "Parameters for deleting a project."
      },
...
      "DeleteUserInput": {
        "properties": {
          "user_email": {
            "type": "string",
            "format": "email",
            "title": "User Email",
            "description": "Email of the user to delete"
          }
        },
        "type": "object",
        "required": [
          "user_email"
        ],
        "title": "DeleteUserInput",
        "description": "Parameters for deleting a user."
      },
      "GetAllProjectsInput": {
        "properties": {
          "user_email": {
            "type": "string",
            "format": "email",
            "title": "User Email",
            "description": "Email of the user to retrieve projects for"
          }
        },
        "type": "object",
        "required": [
          "user_email"
        ],
        "title": "GetAllProjectsInput",
        "description": "Parameters for getting all projects of a user."
      },
      "GetAllProposalsInput": {
        "properties": {
          "user_email": {
            "type": "string",
            "format": "email",
            "title": "User Email",
            "description": "Email of the user"
          },
          "project_name": {
            "type": "string",
            "title": "Project Name",
            "description": "Name of the project"
          }
        },
        "type": "object",
        "required": [
          "user_email",
          "project_name"
        ],
        "title": "GetAllProposalsInput",
        "description": "Parameters for getting all proposals of a project."
      },
...
      "GetProposalInput": {
        "properties": {
          "user_email": {
            "type": "string",
            "format": "email",
            "title": "User Email",
            "description": "Email of the user"
          },
          "project_name": {
            "type": "string",
            "title": "Project Name",
            "description": "Name of the project"
          },
          "proposal_id": {
            "$ref": "#/components/schemas/PydanticObjectId",
            "description": "ID of the proposal to retrieve"
          }
        },
        "type": "object",
        "required": [
          "user_email",
          "project_name",
          "proposal_id"
        ],
        "title": "GetProposalInput",
        "description": "Parameters for getting a proposal."
      },
...
      "GetUserInput": {
        "properties": {
          "user_email": {
            "type": "string",
            "format": "email",
            "title": "User Email",
            "description": "Email of the user to retrieve"
          }
        },
        "type": "object",
        "required": [
          "user_email"
        ],
        "title": "GetUserInput",
        "description": "Parameters for getting a user."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NewProjectInput": {
        "properties": {
          "user_email": {
            "type": "string",
            "format": "email",
            "title": "User Email",
            "description": "Email of the user"
          },
          "project_name": {
            "type": "string",
            "title": "Project Name",
            "description": "Name of the new project"
          },
          "packs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Packs",
            "description": "List of pack names associated with the project"
          }
        },
        "type": "object",
        "required": [
          "user_email",
          "project_name",
          "packs"
        ],
        "title": "NewProjectInput",
        "description": "Parameters for creating a new project."
      },
      "NewUserInput": {
        "properties": {
          "user_name": {
            "type": "string",
            "title": "User Name",
            "description": "Name of the new user"
          },
          "user_email": {
            "type": "string",
            "format": "email",
            "title": "User Email",
            "description": "Email of the new user"
          }
        },
        "type": "object",
        "required": [
          "user_name",
          "user_email"
        ],
        "title": "NewUserInput",
        "description": "Parameters for creating a new user."
      },
...
      "User": {
        "properties": {
          "_id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PydanticObjectId"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Name of the user"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Email of the user"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/Project"
            },
            "type": "array",
            "title": "Projects",
            "description": "List of projects associated with the user"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email"
        ],
        "title": "User",
        "description": "User schema."
      },
...
    }
  }
}